Gain option is expressed in decibels, regardless of receiver type

+++ rtl.h
@@ -5,5 +5,5 @@
 #define RTL_OVERSAMPLE 10
 #define RTL_RATE (SYMBOL_RATE * SPS * RTL_OVERSAMPLE)
 // rtl.c
-void rtl_init(vdl2_state_t *ctx, uint32_t device, int freq, int gain, int correction);
+void rtl_init(vdl2_state_t *ctx, uint32_t device, int freq, float gain, int correction);
 void rtl_cancel();
